fix(storagenode): handle unsealing of already appendable log streams
Ensure that unsealing log streams that are already in an appendable state does
not affect the log stream status. Unsealing changes the state of the log stream
to accept Append RPC and creates new replicate channels when it is the primary
replica. Unsealing an already appendable log stream can be harmful since it
creates additional replicate channels even though replicate channels already
exist. This PR addresses this situation.