feat(storage): add separate sync/wal options for data and commit DBs
This PR introduces flags to disable WAL and sync for data and commit databases:

- `--storage-datadb-disable-wal`
- `--storage-datadb-no-sync`
- `--storage-commitdb-disable-wal`
- `--storage-commitdb-no-sync`

Previously, a single setting existed for WAL and sync, which was applied to both
data and commit databases in storage. This PR makes them more flexible by
allowing separate settings.

Note that turning off sync is efficient for I/O performance, but it can result
in [data loss on process termination](https://github.com/cockroachdb/pebble/issues/2624),
as data will be written into an in-memory buffer for WAL before being written to
OS pages.