perf(reportCollector): reduce calling sampleTracer.report
sampleTracer.report is called as many times as UncommittedLLSNLength
and filtered by sample rate. Normally, there is no problem, but if
the commit processing is delayed in varlogsn, a report with a large
UncommittedLLSNLength may come. In this case, the load on
reportCollector increases. Improved so that it can be called at each
sample rate.