test: fix flaky TestReportCommit
This change fixes a flaky test issue in `TestReportCommit`. The test
occasionally failed with an "index out of range" error because the `GetReport`
request was processed by the mock storage node after the last `CommitBatch` in
`reportsCommits`.
This was not an issue with the storage node or metadata repository but a problem
with the test logic.

Example failure log:

```txt
2025-05-02T06:55:51.0691510Z panic: runtime error: index out of range [5] with length 5
2025-05-02T06:55:51.0691890Z
2025-05-02T06:55:51.0692238Z goroutine 31718 [running]:
2025-05-02T06:55:51.0692988Z github.com/kakao/varlog/internal/metarepos_test.TestReportCommit.func2({0x1da8510, 0xc000120bf0})
```

The fix ensures that the test logic handles cases where the version index
exceeds the available range by using the last available report.