feat(storagenode): enable batching of writes in primary/backup writer
This change introduces batching of write tasks in primary and backup writers.
The number of tasks is limited by `WithMaxWriteTaskBatchLength`. Since each
write task can already have batched writes, its value is limited from 1 to 3.
This is an internal configuration, thus not exposed as a CLI flag. This commit
sets `WithMaxWriteTaskBatchLength` to 3. We will observe this further and can
change it.