feat(storage): use latest Pebble SST format (FormatVirtualSSTables)
This commit updates Pebble to use the latest FormatMajorVersion,
FormatVirtualSSTables, in preparation for upgrading to Pebble v2.

Upgrading FormatMajorVersion in Pebble is transparent, so we do not need manual
updates to the storage layer.

Note:
Downgrading to the previous FormatMajorVersion is not possible. Deploy this
change carefully to existing running clusters.