refactor(storage): rename dataStore to valueStore for clarity
This change renames dataStore to valueStore to clarify its role: it stores user
log entry values. In contrast, commitStore stores the GLSN
(GlobalLogSequenceNumber), which is issued by the metadata repository. The GLSN
acts as the key for the log entry value. While the commitStore could be called
keyStore, we retain the name "commitStore" because it stores global state as
well as GLSNs.

This commit does not change any user-facing features such as the CLI or
environments.