revert(storage): revert Pebble to v1 and its SST format
This reverts commit f95ca6f3 (from #1114) and e8386fb9 (from #1112).

We are downgrading Pebble from v2 back to v1 and reverting the SST file format
from `FormatVirtualSSTables` to the v1 default.

The primary reasons for this rollback are:

- The stability and performance benefits of Pebble v2 have not yet been
  sufficiently verified in the Varlog environment.
- The `FormatVirtualSSTables` format, introduced in #1114, creates a
  backward-incompatible on-disk format. Without confirmed advantages,
  maintaining compatibility with existing clusters is a higher priority.

To ensure the stability of the system, we will temporarily maintain the
dependency on the proven Pebble v1. A more thorough investigation and testing of
Pebble v2 and its new features will be conducted before attempting to upgrade
again in the future.

Refs:

- https://github.com/cockroachdb/pebble/issues/4872
- https://github.com/cockroachdb/pebble/pull/4947