feat(varlog): add observer to subscribe API for metrics
This change introduces an observer to the Subscribe API, allowing users to
monitor internal latencies and diagnose performance issues. Previously, the
internal pipeline of the subscribe client was a black box, making it difficult
to pinpoint the cause of performance issues.

With this feature, users can implement the `SubscribeObserver` interface and
register it using the `WithSubscribeObserver` option. The `Observe` method will
be called for each log entry, providing a `SubscribeStats` struct that contains
a detailed breakdown of the time spent in different stages of the internal
pipeline.

The following statistics are provided:

- TransmitEnqueueDuration: The time taken to enqueue a subscribed log entry into
  the internal transmit queue.
- TransmitQueueWait: The time a log entry waits in the transmit queue, which can
  indicate scheduling delays or high CPU load if the value is high.
- DispatchQueueWait: The time a log entry waits in the dispatch queue before
  being passed to the user's callback.
- ProcessDuration: The execution time of the user-provided callback for a single
  log entry.