refactor(proto): remove deprecated error field from AppendResult
This commit removes the deprecated `error` field from the `AppendResult` message
in `proto/snpb/log_io.proto`.

The field was originally deprecated in #985 as part of the transition to atomic
append operations. With all usage of the field already removed, this commit
completes the cleanup by deleting the field definition entirely.